+++ src/translatable.constants.ts
@@ -1,2 +1,3 @@
 export const TRANSLATABLE_OPTIONS = "TRANSLATABLE_OPTIONS";
 export const TRANSLATABLE_METADATA_KEY = "translatable:fields";
+export const SKIP_TRANSLATION_KEY = "translatable:skip";